This is a sloppy commit. I have done too much work between this one and the last.
I am working towards refactoring the backend of the compiler such that it does not
use a second copy of all of the data structures tailored for the target assembly language.
In my mind, the ideal is to translate the abstract Bytecode directly to the target assembly,
such that we only need to write a new translation for each target assembly. This shouldn't
require the whole gammut of types in order to accomplish. So I have defined an abstract
local register allocator such that locals can be allocated abstractly, and hopefully this
allows for a mapping to just map the abstract registers and stack slots, to actual registers
and stack slots

in addition, I have been removing the Result<T, E> a.k.a Expected<T, E> types from the compiler
in favor of returning a boolean success/failure and using explicit output parameters.
I feel that this API is cleaner, and gives the same functionality.
The error type is stored in the context itself. so there is no global state (like an errno or something)
and it allows for returning an enumeration instead of just a bool, allowing for more nuance when
handling failure, when needed.

+++ exp/include/backend/lifetimes.h
@@ -43,8 +43,8 @@ Lifetime lifetime_immortal();
  * function.
  */
 typedef struct Lifetimes {
-    u64 count;
-    u64 capacity;
+    u32 count;
+    u32 capacity;
     Lifetime *buffer;
 } Lifetimes;
 
@@ -52,8 +52,8 @@ void lifetimes_destroy(Lifetimes *restrict li);
 /**
  * @brief update the lifetime information of the given ssa
  */
-void lifetimes_update(Lifetimes *lifetimes, u64 ssa, Lifetime lifetime);
-Lifetime *lifetimes_at(Lifetimes *restrict li, u64 ssa);
+void lifetimes_update(Lifetimes *lifetimes, u32 ssa, Lifetime lifetime);
+Lifetime *lifetimes_at(Lifetimes *restrict li, u32 ssa);
 
 struct x64_Context;
 void lifetimes_initialize(Lifetimes *lifetimes,
